requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
python-whois
dnspython
requests
